# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Brazilian Artisanal Bags E-commerce API
  description: API documentation for the Brazilian Artisanal Bags E-commerce platform
  version: 0.1.0
  contact:
    email: backend@artesanato-ecommerce.com
servers:
  - url: https://artesanato-ecommerce.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /cart:
    get:
      summary: Get cart items
      description: Returns the current shopping cart with product details
      tags:
        - cart
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add items to cart
      description: Adds or updates items in the shopping cart
      tags:
        - cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/CartItemInput'
      responses:
        '200':
          description: Items added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItemInput'
        '400':
          description: Invalid request or insufficient inventory
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        product_id:
                          type: string
                          format: uuid
                        requested:
                          type: integer
                        available:
                          type: integer
                        error:
                          type: string
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        inventory_count:
          type: integer
        category:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
...
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt_text:
          type: string
        is_primary:
          type: boolean
    CartItem:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        product:
          $ref: '#/components/schemas/Product'
    CartItemInput:
      type: object
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
      required:
        - product_id
        - quantity
...
    Error:
      type: object
      properties:
        error:
          type: string